clean up tests and modularize
