+++ jest-tests/test2.test.js
@@ -1,16 +1,13 @@
 const Parser = require("..");
+const constants = require("./constants");
 const Javascript = require("tree-sitter-javascript");
 const jsParser = new Parser();
 jsParser.setLanguage(Javascript);
 
 describe("Jest test 1", () => {
   it("should work", () => {
-    const code = jsParser.parse(`
-      const Parser = require(".");
-      const Javascript = require("tree-sitter-javascript");
-      const jsParser = new Parser();
-    `)
+    const code = jsParser.parse(constants.INPUT)
     const output = code.rootNode.toString()
-    expect(output).toBe('(program (lexical_declaration (variable_declarator name: (identifier) value: (call_expression function: (identifier) arguments: (arguments (string))))) (lexical_declaration (variable_declarator name: (identifier) value: (call_expression function: (identifier) arguments: (arguments (string))))) (lexical_declaration (variable_declarator name: (identifier) value: (new_expression constructor: (identifier) arguments: (arguments)))))');
+    expect(output).toBe(constants.OUTPUT);
   })
 })
